08-19
完成了充值的账户信息添加逻辑
页面充值加上了通道判断逻辑

+++ weizu-parent/weizu-web-modules/weizu-channel/src/main/java/com/weizu/flowsys/web/agency/dao/impl/ChargeRecordDao.java
@@ -14,7 +14,7 @@ import com.weizu.flowsys.web.agency.pojo.ChargeRecordPo;
 import com.weizu.flowsys.web.agency.pojo.ConsumeRecordPo;
 
 @Repository("chargeRecordDao")
-public class ChargeRecordDao extends DaoImpl<ChargeRecordPo, Integer> implements ChargeRecordDaoInterface {
+public class ChargeRecordDao extends DaoImpl<ChargeRecordPo, Long> implements ChargeRecordDaoInterface {
 
 	@Resource
 	private SqlSessionTemplate sqlSessionTemplate;
